# openapi-specification.yaml
openapi: 3.0.3
info:
  title: E-Bookstore API
  description: Backend API for an online bookstore application
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local development server

tags:
  - name: Authentication
    description: User authentication operations
  - name: Products
    description: Product catalog operations
  - name: Cart
    description: Shopping cart operations
  - name: Orders
    description: Order management operations
  - name: Users
    description: User profile operations

paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'

  /products:
    get:
      tags:
        - Products
      summary: Get all products with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: brand
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'

    post:
      tags:
        - Products
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Product created successfully

  /products/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'

  /cart:
    get:
      tags:
        - Cart
      summary: Get user's cart
      responses:
        '200':
          description: Cart details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'

  /cart/items:
    post:
      tags:
        - Cart
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItemRequest'
      responses:
        '201':
          description: Item added to cart

  /cart/items/{id}:
    put:
      tags:
        - Cart
      summary: Update cart item quantity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
      responses:
        '200':
          description: Cart item updated

    delete:
      tags:
        - Cart
      summary: Remove item from cart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Item removed from cart

  /orders:
    get:
      tags:
        - Orders
      summary: Get user's order history
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'

    post:
      tags:
        - Orders
      summary: Create new order (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

  /orders/{id}/cancel:
    post:
      tags:
        - Orders
      summary: Cancel order (within 48 hours)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Order cancelled successfully

  /users/profile:
    get:
      tags:
        - Users
      summary: Get user profile
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

components:
  schemas:
    UserRegistration:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        name:
          type: string
        phone:
          type: string

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string

    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'

    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        name:
          type: string
        phone:
          type: string
        giftPoints:
          type: integer
        createdAt:
          type: string
          format: date-time

    ProductRequest:
      type: object
      required:
        - title
        - author
        - price
        - category
      properties:
        title:
          type: string
        author:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        category:
          type: string
        brand:
          type: string
        stockQuantity:
          type: integer
        imageUrl:
          type: string

    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        author:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        category:
          type: string
        brand:
          type: string
        stockQuantity:
          type: integer
        imageUrl:
          type: string
        createdAt:
          type: string
          format: date-time

    CartItemRequest:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: integer
          format: int64
        quantity:
          type: integer

    CartItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        product:
          $ref: '#/components/schemas/Product'
        quantity:
          type: integer
        subtotal:
          type: number
          format: double

    Cart:
      type: object
      properties:
        id:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        total:
          type: number
          format: double

    OrderRequest:
      type: object
      required:
        - deliveryAddress
        - paymentMethod
      properties:
        deliveryAddress:
          type: string
        paymentMethod:
          type: string
          enum: [CREDIT_CARD, DEBIT_CARD]
        useGiftPoints:
          type: integer
          default: 0

    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/User'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalAmount:
          type: number
          format: double
        deliveryAddress:
          type: string
        paymentMethod:
          type: string
        status:
          type: string
          enum: [PENDING, CONFIRMED, SHIPPED, DELIVERED, CANCELLED]
        giftPointsUsed:
          type: integer
        orderDate:
          type: string
          format: date-time

    OrderItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        product:
          $ref: '#/components/schemas/Product'
        quantity:
          type: integer
        price:
          type: number
          format: double

# Made with Bob